social_media_api/requirements.txt
Django==4.2.16
djangorestframework==3.14.0
django-cors-headers==4.3.1
Pillow==10.2.0
